Remove the option to pass output files inline.
This was a premature optimization in the original protocol design.
Generally speaking, the server does not know whether or not the client
is interested in an output file. A client using the output file only as
an intermediate result to be fed back into a later action never needs to
download it, and instead can just pass the digest into the next action.
Bazel does not implement this behaviour because of architectural
constraints at the matter, but it is desired.

The same is not being done for stdout/stderr logs. The client is much
more reliably interested in them, especially for failing actions. In
fact, it was the proposal to include them inline that was extended to
output files.