Add a working directory to the Action message.
This allows a job to be started from a directory other than the input
root. While Bazel does not need this functionality, it is not uncommon
for other build systems to descend into subdirectories to run their
actions, and forcing them to either re-architect so that they run from
input root or include a wrapper script is not optimal, especially given
that such wrappers are platform-dependent and, on platforms with high
process startup latency, may cost performance.