Allowing file symlink outputs. (#28)
* Allowing file symlink outputs.

This is a missing case from the existing API: symlink output directories
are already supported, because the SymlinkNodes are part of the
Directory message. Similarly, all file symlinks that are part of larger
output directories are already supported as well. However, if the output
is a file, then the current API does not distinguish between it being a
symlink or not, and instead just provides the content Digest.

A motivating case for distinguishing the symlink case is efficiency,
when the client did not want the action to create and download a copy of
an existing file.

TESTED=bazel build

	modified:   build/bazel/remote/execution/v2/remote_execution.proto

* Adding output_directory_symlinks field.

* Fixing OutputSymlink comment.